refactor: split oversized server files into module directories (#625)
Split 4 files that exceeded the 800-line limit into module directories,
preserving all public import paths via pub use re-exports. No behavior changes.

- task_db.rs (1933 lines) → task_db/{mod,schema,artifacts,checkpoints,queries,row}.rs
- task_runner.rs (3097 lines) → task_runner/{mod,types,store,scheduling}.rs
- http.rs (2503 lines) → http/mod.rs (already had http/ subdirectory for submodules)
- periodic_reviewer.rs (2096 lines) → periodic_reviewer/mod.rs
