+++ crates/harness-server/src/periodic_reviewer/mod.rs
@@ -1692,10 +1692,10 @@ mod tests {
         // The forbidden pattern is split across two literals so that this very
         // test does not trigger the assertion it is enforcing.
         let forbidden = ["Command::new(", "\"git\")"].concat();
-        let source = include_str!("periodic_reviewer.rs");
+        let source = include_str!("mod.rs");
         assert!(
             !source.contains(&forbidden),
-            "periodic_reviewer.rs must not spawn git directly"
+            "periodic_reviewer/mod.rs must not spawn git directly"
         );
     }
