+++ crates/harness-server/src/task_db/mod.rs
@@ -1,186 +1,25 @@
-use crate::http::parse_pr_num_from_url;
-use crate::task_runner::{TaskState, TaskStatus};
-use harness_core::db::{open_pool, Migration, Migrator};
-use harness_core::error::TaskDbDecodeError;
-use serde::{Deserialize, Serialize};
+mod artifacts;
+mod checkpoints;
+mod queries;
+mod row;
+mod schema;
+
+pub use artifacts::TaskArtifact;
+pub use checkpoints::{RecoveryResult, TaskCheckpoint};
+
+use harness_core::db::{open_pool, Migrator};
+use schema::TASK_MIGRATIONS;
 use sqlx::sqlite::SqlitePool;
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
-/// Maximum artifact content size in bytes before truncation.
-const ARTIFACT_MAX_BYTES: usize = 65_536;
-
-/// Column list for `query_as::<_, TaskRow>` — single source of truth.
-///
-/// Every SELECT that maps to `TaskRow` MUST use this constant.
-/// When adding a field to `TaskRow`, add the column here once and all queries
-/// pick it up automatically.  The `task_row_columns_match_struct` test below
-/// will fail if this list drifts from the struct definition.
-const TASK_ROW_COLUMNS: &str = "id, status, turn, pr_url, rounds, error, source, external_id, parent_id, created_at, repo, depends_on, project, priority, phase, description";
-
-/// Versioned migrations for the tasks table.
-///
-/// v1 – baseline schema (all columns including those added in later iterations)
-/// v2/v3/v4 – additive ALTER TABLE for databases that predate v1 tracking;
-///   duplicate-column errors are silently ignored by the Migrator.
-/// v5 – add task_artifacts table for persisting agent output per task turn.
-/// v10 – add composite index on (project, status, updated_at).
-/// v11 – add task_checkpoints table for phase recovery.
-/// v12 – add priority column for priority-based task scheduling.
-/// v13 – add phase column for consistent phase persistence across cache/DB paths.
-/// v14 – add description column for task observability after restart.
-static TASK_MIGRATIONS: &[Migration] = &[
-    Migration {
-        version: 1,
-        description: "create tasks table",
-        sql: "CREATE TABLE IF NOT EXISTS tasks (
-            id          TEXT PRIMARY KEY,
-            status      TEXT NOT NULL DEFAULT 'pending',
-            turn        INTEGER NOT NULL DEFAULT 0,
-            pr_url      TEXT,
-            rounds      TEXT NOT NULL DEFAULT '[]',
-            error       TEXT,
-            created_at  TEXT NOT NULL DEFAULT (datetime('now')),
-            updated_at  TEXT NOT NULL DEFAULT (datetime('now'))
-        )",
-    },
-    Migration {
-        version: 2,
-        description: "add source column",
-        sql: "ALTER TABLE tasks ADD COLUMN source TEXT",
-    },
-    Migration {
-        version: 3,
-        description: "add external_id column",
-        sql: "ALTER TABLE tasks ADD COLUMN external_id TEXT",
-    },
-    Migration {
-        version: 4,
-        description: "add parent_id column",
-        sql: "ALTER TABLE tasks ADD COLUMN parent_id TEXT",
-    },
-    Migration {
-        version: 5,
-        description: "create task_artifacts table",
-        sql: "CREATE TABLE IF NOT EXISTS task_artifacts (
-            id            INTEGER PRIMARY KEY AUTOINCREMENT,
-            task_id       TEXT NOT NULL,
-            turn          INTEGER NOT NULL DEFAULT 0,
-            artifact_type TEXT NOT NULL,
-            content       TEXT NOT NULL,
-            created_at    TEXT NOT NULL DEFAULT (datetime('now'))
-        )",
-    },
-    Migration {
-        version: 6,
-        description: "add repo column",
-        sql: "ALTER TABLE tasks ADD COLUMN repo TEXT",
-    },
-    Migration {
-        version: 7,
-        description: "add depends_on column",
-        sql: "ALTER TABLE tasks ADD COLUMN depends_on TEXT NOT NULL DEFAULT '[]'",
-    },
-    Migration {
-        version: 8,
-        description: "add project column for task observability",
-        sql: "ALTER TABLE tasks ADD COLUMN project TEXT",
-    },
-    Migration {
-        version: 10,
-        description: "add index on tasks(project, status, updated_at) for dashboard queries",
-        sql: "CREATE INDEX IF NOT EXISTS idx_tasks_project_status_updated \
-              ON tasks(project, status, updated_at DESC)",
-    },
-    Migration {
-        version: 11,
-        description: "create task_checkpoints table for phase recovery",
-        sql: "CREATE TABLE IF NOT EXISTS task_checkpoints (
-            task_id       TEXT PRIMARY KEY,
-            triage_output TEXT,
-            plan_output   TEXT,
-            pr_url        TEXT,
-            last_phase    TEXT NOT NULL,
-            updated_at    TEXT NOT NULL
-        )",
-    },
-    Migration {
-        version: 12,
-        description: "add priority column for task scheduling",
-        sql: "ALTER TABLE tasks ADD COLUMN priority INTEGER NOT NULL DEFAULT 0",
-    },
-    Migration {
-        version: 13,
-        description: "add phase column for consistent phase persistence",
-        sql: r#"ALTER TABLE tasks ADD COLUMN phase TEXT NOT NULL DEFAULT '"implement"'"#,
-    },
-    Migration {
-        version: 14,
-        description: "add description column for task observability",
-        sql: "ALTER TABLE tasks ADD COLUMN description TEXT",
-    },
-];
-
-/// A single persisted artifact captured from agent output during task execution.
-#[derive(Debug, Clone, Serialize, Deserialize, sqlx::FromRow)]
-pub struct TaskArtifact {
-    pub task_id: String,
-    pub turn: i64,
-    pub artifact_type: String,
-    pub content: String,
-    pub created_at: String,
-}
-
-/// Persisted phase checkpoint for a task, used to resume after server restart.
-///
-/// A single row per task (upsert semantics). Each field is populated once the
-/// corresponding phase completes; earlier fields are preserved on subsequent writes.
-#[derive(Debug, Clone, sqlx::FromRow)]
-pub struct TaskCheckpoint {
-    pub task_id: String,
-    /// Non-null once the Triage phase has completed.
-    pub triage_output: Option<String>,
-    /// Non-null once the Plan phase has completed.
-    pub plan_output: Option<String>,
-    /// Non-null once a PR has been created (most critical for duplicate-PR prevention).
-    pub pr_url: Option<String>,
-    /// Last completed phase: `triage_done` | `plan_done` | `pr_created`.
-    pub last_phase: String,
-    pub updated_at: String,
-}
-
-/// Result of [`TaskDb::recover_in_progress`].
-#[derive(Debug, Default)]
-pub struct RecoveryResult {
-    /// Tasks that were in interrupted states and are now `failed`.
-    pub failed: u32,
-    /// Tasks that were in interrupted states and have checkpoints — now `pending` for resume.
-    pub resumed: u32,
-    /// Tasks that were `pending` mid-transient-retry at crash time and are now `failed`.
-    pub transient_failed: u32,
-}
-
-/// Row returned by the checkpoint JOIN query inside [`TaskDb::recover_in_progress`].
-#[derive(sqlx::FromRow)]
-struct RecoveryRow {
-    id: String,
-    status: String,
-    turn: i64,
-    task_pr_url: Option<String>,
-    triage_output: Option<String>,
-    plan_output: Option<String>,
-    ck_pr_url: Option<String>,
-}
+use crate::task_runner::TaskState;
 
 pub struct TaskDb {
     pool: SqlitePool,
 }
 
 impl TaskDb {
-    fn status_placeholders(len: usize) -> String {
-        std::iter::repeat_n("?", len).collect::<Vec<_>>().join(", ")
-    }
-
     pub async fn open(path: &Path) -> anyhow::Result<Self> {
         let pool = open_pool(path).await?;
         let db = Self { pool };
@@ -189,440 +28,74 @@ impl TaskDb {
     }
 
     pub async fn insert(&self, state: &TaskState) -> anyhow::Result<()> {
-        let rounds_json = serde_json::to_string(&state.rounds)?;
-        let depends_on_json = serde_json::to_string(&state.depends_on)?;
-        let status = state.status.as_ref();
-        let phase_json = serde_json::to_string(&state.phase)?;
-        sqlx::query(
-            "INSERT INTO tasks (id, status, turn, pr_url, rounds, error, source, external_id, parent_id, created_at, repo, depends_on, project, priority, phase, description)
-             VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, COALESCE(?, datetime('now')), ?, ?, ?, ?, ?, ?)",
-        )
-        .bind(&state.id.0)
-        .bind(status)
-        .bind(state.turn as i64)
-        .bind(&state.pr_url)
-        .bind(&rounds_json)
-        .bind(&state.error)
-        .bind(&state.source)
-        .bind(&state.external_id)
-        .bind(state.parent_id.as_ref().map(|id| &id.0))
-        .bind(&state.created_at)
-        .bind(&state.repo)
-        .bind(&depends_on_json)
-        .bind(state.project_root.as_ref().map(|p| p.to_string_lossy().into_owned()))
-        .bind(state.priority as i64)
-        .bind(&phase_json)
-        .bind(state.description.as_deref())
-        .execute(&self.pool)
-        .await?;
-        Ok(())
+        queries::insert(&self.pool, state).await
     }
 
     pub async fn update(&self, state: &TaskState) -> anyhow::Result<()> {
-        let rounds_json = serde_json::to_string(&state.rounds)?;
-        let depends_on_json = serde_json::to_string(&state.depends_on)?;
-        let phase_json = serde_json::to_string(&state.phase)?;
-        let status = state.status.as_ref();
-        sqlx::query(
-            "UPDATE tasks SET status = ?, turn = ?, pr_url = ?, rounds = ?, error = ?,
-                    source = ?, external_id = ?, repo = ?, depends_on = ?, project = ?,
-                    priority = ?, phase = ?, description = ?, updated_at = datetime('now')
-             WHERE id = ?",
-        )
-        .bind(status)
-        .bind(state.turn as i64)
-        .bind(&state.pr_url)
-        .bind(&rounds_json)
-        .bind(&state.error)
-        .bind(&state.source)
-        .bind(&state.external_id)
-        .bind(&state.repo)
-        .bind(&depends_on_json)
-        .bind(
-            state
-                .project_root
-                .as_ref()
-                .map(|p| p.to_string_lossy().into_owned()),
-        )
-        .bind(state.priority as i64)
-        .bind(&phase_json)
-        .bind(state.description.as_deref())
-        .bind(&state.id.0)
-        .execute(&self.pool)
-        .await?;
-        Ok(())
+        queries::update(&self.pool, state).await
     }
 
     pub async fn get(&self, id: &str) -> anyhow::Result<Option<TaskState>> {
-        let sql = format!("SELECT {TASK_ROW_COLUMNS} FROM tasks WHERE id = ?");
-        let row = sqlx::query_as::<_, TaskRow>(&sql)
-            .bind(id)
-            .fetch_optional(&self.pool)
-            .await?;
-        row.map(TaskRow::try_into_task_state).transpose()
+        queries::get(&self.pool, id).await
     }
 
     pub async fn list(&self) -> anyhow::Result<Vec<TaskState>> {
-        let sql = format!("SELECT {TASK_ROW_COLUMNS} FROM tasks ORDER BY created_at DESC");
-        let rows = sqlx::query_as::<_, TaskRow>(&sql)
-            .fetch_all(&self.pool)
-            .await?;
-        rows.into_iter().map(TaskRow::try_into_task_state).collect()
+        queries::list(&self.pool).await
     }
 
-    /// Return tasks whose `status` column matches any value in `statuses`.
-    ///
-    /// Uses parameterized placeholders — safe for internal status string constants.
-    /// Returns an empty `Vec` when `statuses` is empty.
     pub async fn list_by_status(&self, statuses: &[&str]) -> anyhow::Result<Vec<TaskState>> {
-        if statuses.is_empty() {
-            return Ok(Vec::new());
-        }
-        let placeholders = Self::status_placeholders(statuses.len());
-        let sql = format!(
-            "SELECT {TASK_ROW_COLUMNS} FROM tasks WHERE status IN ({placeholders}) ORDER BY created_at DESC"
-        );
-        let mut q = sqlx::query_as::<_, TaskRow>(&sql);
-        for status in statuses {
-            q = q.bind(*status);
-        }
-        let rows = q.fetch_all(&self.pool).await?;
-        rows.into_iter().map(TaskRow::try_into_task_state).collect()
+        queries::list_by_status(&self.pool, statuses).await
     }
 
-    /// Find an active (non-terminal) task for the same project + external_id.
     pub async fn find_active_duplicate(
         &self,
         project: &str,
         external_id: &str,
     ) -> anyhow::Result<Option<String>> {
-        let row: Option<(String,)> = sqlx::query_as(
-            "SELECT id FROM tasks WHERE project = ? AND external_id = ? \
-             AND status IN ('pending', 'awaiting_deps', 'implementing', 'agent_review', 'waiting', 'reviewing') \
-             LIMIT 1",
-        )
-        .bind(project)
-        .bind(external_id)
-        .fetch_optional(&self.pool)
-        .await?;
-        Ok(row.map(|r| r.0))
+        queries::find_active_duplicate(&self.pool, project, external_id).await
     }
 
-    /// Find a `done` task for the same project + external_id that has a non-null `pr_url`.
-    ///
-    /// Returns `(task_id, pr_url)` when found. Only matches `done` — failed/cancelled
-    /// tasks are excluded so that retries after failure are always permitted.
     pub async fn find_terminal_with_pr(
         &self,
         project: &str,
         external_id: &str,
     ) -> anyhow::Result<Option<(String, String)>> {
-        let row: Option<(String, String)> = sqlx::query_as(
-            "SELECT id, pr_url FROM tasks \
-             WHERE project = ? AND external_id = ? AND status = 'done' AND pr_url IS NOT NULL \
-             ORDER BY created_at DESC LIMIT 1",
-        )
-        .bind(project)
-        .bind(external_id)
-        .fetch_optional(&self.pool)
-        .await?;
-        Ok(row)
+        queries::find_terminal_with_pr(&self.pool, project, external_id).await
     }
 
-    /// Return all tasks as lightweight summaries, skipping the heavy `rounds` column.
-    ///
-    /// Used by the `/tasks` list endpoint to avoid deserializing large round histories
-    /// when only summary fields are needed.
     pub async fn list_summaries(&self) -> anyhow::Result<Vec<crate::task_runner::TaskSummary>> {
-        let rows = sqlx::query_as::<_, TaskSummaryRow>(
-            "SELECT id, status, turn, pr_url, error, source, external_id, parent_id, \
-             created_at, repo, depends_on, project, phase, description \
-             FROM tasks ORDER BY created_at DESC",
-        )
-        .fetch_all(&self.pool)
-        .await?;
-        let mut summaries = Vec::with_capacity(rows.len());
-        for row in rows {
-            let id = row.id.clone();
-            match TaskSummaryRow::try_into_summary(row) {
-                Ok(summary) => summaries.push(summary),
-                Err(e) => {
-                    tracing::warn!(task_id = %id, "skipping malformed task row in list_summaries: {e}");
-                }
-            }
-        }
-        Ok(summaries)
+        queries::list_summaries(&self.pool).await
     }
 
-    /// Return `(id, status)` pairs for all tasks — skips all heavy columns.
-    ///
-    /// Used by hot-path callers that only need task status for aggregation
-    /// (e.g. skill governance scoring).
     pub async fn list_id_status(&self) -> anyhow::Result<Vec<(String, String)>> {
-        let rows: Vec<(String, String)> =
-            sqlx::query_as("SELECT id, status FROM tasks ORDER BY created_at DESC")
-                .fetch_all(&self.pool)
-                .await?;
-        Ok(rows)
+        queries::list_id_status(&self.pool).await
     }
 
-    /// Return only task IDs whose `status` matches any value in `statuses`.
-    ///
-    /// Skips all heavy columns (rounds, error, etc.) — use this when only IDs
-    /// are needed (e.g. orphan-worktree cleanup at startup).
     pub async fn list_ids_by_status(&self, statuses: &[&str]) -> anyhow::Result<Vec<String>> {
-        if statuses.is_empty() {
-            return Ok(Vec::new());
-        }
-        let placeholders = Self::status_placeholders(statuses.len());
-        let sql = format!("SELECT id FROM tasks WHERE status IN ({placeholders})");
-        let mut q = sqlx::query_as::<_, (String,)>(&sql);
-        for status in statuses {
-            q = q.bind(*status);
-        }
-        let rows = q.fetch_all(&self.pool).await?;
-        Ok(rows.into_iter().map(|(id,)| id).collect())
+        queries::list_ids_by_status(&self.pool, statuses).await
     }
 
-    /// Return the status of a single task, fetching only the `status` column.
-    ///
-    /// Much lighter than `get()` — avoids deserializing the `rounds` JSON.
-    /// Used by `check_awaiting_deps` to resolve dependency status with a single
-    /// DB round-trip instead of a full row fetch.
     pub async fn get_status_only(&self, id: &str) -> anyhow::Result<Option<String>> {
-        let row: Option<(String,)> = sqlx::query_as("SELECT status FROM tasks WHERE id = ?")
-            .bind(id)
-            .fetch_optional(&self.pool)
-            .await?;
-        Ok(row.map(|(s,)| s))
+        queries::get_status_only(&self.pool, id).await
     }
 
-    /// Return `true` if a task row with the given ID exists in the database.
     pub async fn exists_by_id(&self, id: &str) -> anyhow::Result<bool> {
-        let row: Option<(String,)> = sqlx::query_as("SELECT id FROM tasks WHERE id = ?")
-            .bind(id)
-            .fetch_optional(&self.pool)
-            .await?;
-        Ok(row.is_some())
+        queries::exists_by_id(&self.pool, id).await
     }
 
-    /// Apply event-replayed state to a task row.
-    ///
-    /// Called during startup, **before** `recover_in_progress()`, so that
-    /// event-sourced data takes precedence over checkpoint data.
-    ///
-    /// - If `terminal_status` is `Some`, the task's status is overwritten
-    ///   (only while the row is still in an interrupted state).
-    /// - If `pr_url` is `Some` and the row's `pr_url` is currently `NULL`,
-    ///   the value is written back so `recover_in_progress()` can resume the task.
     pub async fn apply_replayed_state(
         &self,
         task_id: &str,
         pr_url: Option<&str>,
         terminal_status: Option<&str>,
     ) -> anyhow::Result<()> {
-        if let Some(status) = terminal_status {
-            // Overwrite to terminal status; only touches tasks still in interrupted states
-            // so we never downgrade a task that already reached Done/Failed in the DB.
-            let sql = format!(
-                "UPDATE tasks SET status = ?, pr_url = COALESCE(?, pr_url), \
-                 updated_at = datetime('now') \
-                 WHERE id = ? \
-                 AND status IN ({})",
-                Self::status_placeholders(TaskStatus::resumable_statuses().len())
-            );
-            let query = sqlx::query(&sql).bind(status).bind(pr_url).bind(task_id);
-            let query = TaskStatus::resumable_statuses()
-                .iter()
-                .fold(query, |query, resumable| query.bind(*resumable));
-            query.execute(&self.pool).await?;
-        } else if let Some(url) = pr_url {
-            // Write pr_url back only when the DB row currently has no pr_url.
-            sqlx::query(
-                "UPDATE tasks SET pr_url = ? \
-                 WHERE id = ? AND pr_url IS NULL",
-            )
-            .bind(url)
-            .bind(task_id)
-            .execute(&self.pool)
-            .await?;
-        }
-        Ok(())
+        queries::apply_replayed_state(&self.pool, task_id, pr_url, terminal_status).await
     }
 
-    /// Recovery on server restart.
-    ///
-    /// For each interrupted task (`implementing`, `agent_review`, `reviewing`, `waiting`),
-    /// checks the `task_checkpoints` table and the existing `tasks.pr_url` column to
-    /// decide between resuming and failing:
-    ///
-    /// - Has PR (tasks.pr_url or checkpoint.pr_url) → `pending`, `resumed += 1`
-    ///   (executor will skip implement and jump to agent review)
-    /// - Has plan checkpoint → `pending`, `resumed += 1`
-    ///   (executor will use saved plan, skip triage/plan pipeline)
-    /// - Has triage checkpoint only → `pending`, `resumed += 1`
-    ///   (executor will re-run from plan phase)
-    /// - No checkpoint → `failed`, `failed += 1`
-    ///   (existing fail-closed behavior, no safe resume point)
-    ///
-    /// Tasks in `AwaitingDeps` are left unchanged — `check_awaiting_deps()` handles them.
-    /// `pending` tasks in transient-retry are failed (crashed during backoff, cannot re-dispatch).
-    ///
-    /// Returns a [`RecoveryResult`] with counts for each outcome.
     pub async fn recover_in_progress(&self) -> anyhow::Result<RecoveryResult> {
-        // Collect all interrupted tasks with their checkpoint data via LEFT JOIN.
-        let rows = {
-            let sql = format!(
-                "SELECT t.id, t.status, t.turn, t.pr_url AS task_pr_url,
-                        c.triage_output, c.plan_output, c.pr_url AS ck_pr_url
-                 FROM tasks t
-                 LEFT JOIN task_checkpoints c ON t.id = c.task_id
-                 WHERE t.status IN ({})",
-                Self::status_placeholders(TaskStatus::resumable_statuses().len())
-            );
-            let query = TaskStatus::resumable_statuses()
-                .iter()
-                .fold(sqlx::query_as::<_, RecoveryRow>(&sql), |query, status| {
-                    query.bind(*status)
-                });
-            query.fetch_all(&self.pool).await?
-        };
-
-        let mut result = RecoveryResult::default();
-
-        for row in rows {
-            // Fall back to ck_pr_url only when task_pr_url is absent *or* unparseable,
-            // so a corrupted task_pr_url does not mask a valid checkpoint URL.
-            let effective_pr_url = row
-                .task_pr_url
-                .as_deref()
-                .filter(|u| parse_pr_num_from_url(u).is_some())
-                .or_else(|| {
-                    row.ck_pr_url
-                        .as_deref()
-                        .filter(|u| parse_pr_num_from_url(u).is_some())
-                });
-            let has_pr = effective_pr_url.is_some();
-            let has_plan = row.plan_output.is_some();
-            let has_triage = row.triage_output.is_some();
-
-            if has_pr || has_plan || has_triage {
-                // Resume: set back to pending with a diagnostic error field.
-                let reason = if has_pr {
-                    format!(
-                        "resumed after restart (was: {}, pr: {})",
-                        row.status,
-                        effective_pr_url.unwrap_or("checkpoint")
-                    )
-                } else if has_plan {
-                    format!(
-                        "resumed after restart (was: {}, plan checkpoint)",
-                        row.status
-                    )
-                } else {
-                    format!(
-                        "resumed after restart (was: {}, triage checkpoint)",
-                        row.status
-                    )
-                };
-
-                // If task_pr_url was absent or unparseable but ck_pr_url is valid,
-                // write the effective URL back to tasks.pr_url so that re-dispatch
-                // (http.rs) and validate_recovered_tasks() (task_runner.rs) — which
-                // both filter on pr_url.is_some() — can actually see the PR.
-                let task_pr_url_valid = row
-                    .task_pr_url
-                    .as_deref()
-                    .map(|u| parse_pr_num_from_url(u).is_some())
-                    .unwrap_or(false);
-                let needs_pr_url_writeback = !task_pr_url_valid && effective_pr_url.is_some();
-
-                if needs_pr_url_writeback {
-                    sqlx::query(
-                        "UPDATE tasks SET status = 'pending', pr_url = ?, error = ?, \
-                         updated_at = datetime('now') WHERE id = ?",
-                    )
-                    .bind(effective_pr_url)
-                    .bind(&reason)
-                    .bind(&row.id)
-                    .execute(&self.pool)
-                    .await?;
-                    tracing::info!(
-                        task_id = %row.id,
-                        was = %row.status,
-                        pr_url = ?effective_pr_url,
-                        "startup recovery: wrote back pr_url from checkpoint"
-                    );
-                } else {
-                    sqlx::query(
-                        "UPDATE tasks SET status = 'pending', error = ?, updated_at = datetime('now') \
-                         WHERE id = ?",
-                    )
-                    .bind(&reason)
-                    .bind(&row.id)
-                    .execute(&self.pool)
-                    .await?;
-                }
-                result.resumed += 1;
-                tracing::info!(
-                    task_id = %row.id,
-                    was = %row.status,
-                    reason = %reason,
-                    "startup recovery: resumed task"
-                );
-            } else {
-                // Fail: no checkpoint, no safe resume point.
-                let err = format!(
-                    "recovered after restart (was: {}, round: {}, pr: {})",
-                    row.status,
-                    row.turn,
-                    row.task_pr_url.as_deref().unwrap_or("none")
-                );
-                sqlx::query(
-                    "UPDATE tasks SET status = 'failed', error = ?, updated_at = datetime('now') \
-                     WHERE id = ?",
-                )
-                .bind(&err)
-                .bind(&row.id)
-                .execute(&self.pool)
-                .await?;
-                result.failed += 1;
-            }
-        }
-
-        // Tasks that were mid-transient-retry (status=pending, error starts with
-        // "retrying after transient failure") crashed during the backoff window.
-        // They have no PR yet and no persisted issue/prompt, so they cannot be
-        // re-dispatched. Mark them failed so they don't silently stay pending forever.
-        let transient_failed = sqlx::query(
-            "UPDATE tasks \
-             SET status = 'failed', \
-                 error = 'recovered after restart (was: pending in transient retry): ' \
-                      || COALESCE(error, ''), \
-                 updated_at = datetime('now') \
-             WHERE status = 'pending' \
-               AND error LIKE 'retrying after transient failure%'",
-        )
-        .execute(&self.pool)
-        .await?
-        .rows_affected() as u32;
-
-        if transient_failed > 0 {
-            tracing::info!(
-                "startup recovery: failed {} task(s) that were pending mid-transient-retry",
-                transient_failed
-            );
-        }
-        result.transient_failed = transient_failed;
-
-        Ok(result)
+        checkpoints::recover_in_progress(&self.pool).await
     }
 
-    /// Upsert a phase checkpoint for the given task.
-    ///
-    /// Each call advances `last_phase` and updates `updated_at`. Previously saved fields
-    /// (`triage_output`, `plan_output`, `pr_url`) are preserved via `COALESCE` when the
-    /// incoming value is `NULL`, so callers only need to pass the newly available field.
     pub async fn write_checkpoint(
         &self,
         task_id: &str,
@@ -631,150 +104,48 @@ impl TaskDb {
         pr_url: Option<&str>,
         last_phase: &str,
     ) -> anyhow::Result<()> {
-        sqlx::query(
-            "INSERT INTO task_checkpoints \
-                 (task_id, triage_output, plan_output, pr_url, last_phase, updated_at) \
-             VALUES (?, ?, ?, ?, ?, datetime('now')) \
-             ON CONFLICT(task_id) DO UPDATE SET \
-                 triage_output = COALESCE(excluded.triage_output, task_checkpoints.triage_output), \
-                 plan_output   = COALESCE(excluded.plan_output,   task_checkpoints.plan_output), \
-                 pr_url        = COALESCE(excluded.pr_url,        task_checkpoints.pr_url), \
-                 last_phase    = excluded.last_phase, \
-                 updated_at    = excluded.updated_at",
+        checkpoints::write_checkpoint(
+            &self.pool,
+            task_id,
+            triage_output,
+            plan_output,
+            pr_url,
+            last_phase,
         )
-        .bind(task_id)
-        .bind(triage_output)
-        .bind(plan_output)
-        .bind(pr_url)
-        .bind(last_phase)
-        .execute(&self.pool)
-        .await?;
-        Ok(())
+        .await
     }
 
-    /// Load the checkpoint for `task_id`, or `None` if no checkpoint exists.
     pub async fn load_checkpoint(&self, task_id: &str) -> anyhow::Result<Option<TaskCheckpoint>> {
-        let row = sqlx::query_as::<_, TaskCheckpoint>(
-            "SELECT task_id, triage_output, plan_output, pr_url, last_phase, updated_at \
-             FROM task_checkpoints WHERE task_id = ?",
-        )
-        .bind(task_id)
-        .fetch_optional(&self.pool)
-        .await?;
-        Ok(row)
+        checkpoints::load_checkpoint(&self.pool, task_id).await
     }
 
-    /// Return the `pr_url` of the most recently completed Done task that has one, or `None`.
-    /// Orders by `updated_at DESC` because `updated_at` is written when the task transitions
-    /// to Done, which correctly reflects completion time rather than creation time.
     pub async fn latest_done_pr_url(&self) -> anyhow::Result<Option<String>> {
-        let row: Option<(Option<String>,)> = sqlx::query_as(
-            "SELECT pr_url FROM tasks WHERE status = 'done' AND pr_url IS NOT NULL \
-             ORDER BY updated_at DESC LIMIT 1",
-        )
-        .fetch_optional(&self.pool)
-        .await?;
-        Ok(row.and_then(|(pr_url,)| pr_url))
+        queries::latest_done_pr_url(&self.pool).await
     }
 
-    /// Return the `pr_url` of the most recent Done task for the given project root path.
     pub async fn latest_done_pr_url_by_project(
         &self,
         project: &str,
     ) -> anyhow::Result<Option<String>> {
-        let row: Option<(Option<String>,)> = sqlx::query_as(
-            "SELECT pr_url FROM tasks \
-             WHERE status = 'done' AND pr_url IS NOT NULL AND project = ?1 \
-             ORDER BY updated_at DESC LIMIT 1",
-        )
-        .bind(project)
-        .fetch_optional(&self.pool)
-        .await?;
-        Ok(row.and_then(|(pr_url,)| pr_url))
+        queries::latest_done_pr_url_by_project(&self.pool, project).await
     }
 
-    /// Return the latest done PR URL for every project that has one, in a single query.
-    /// The map key is the project root path string; the value is the PR URL.
     pub async fn latest_done_pr_urls_all_projects(
         &self,
     ) -> anyhow::Result<HashMap<String, String>> {
-        let rows: Vec<(String, String)> = sqlx::query_as(
-            "SELECT project, pr_url FROM (\
-               SELECT project, pr_url, \
-                      ROW_NUMBER() OVER (PARTITION BY project ORDER BY updated_at DESC) AS rn \
-               FROM tasks \
-               WHERE status = 'done' AND pr_url IS NOT NULL AND project IS NOT NULL\
-             ) WHERE rn = 1",
-        )
-        .fetch_all(&self.pool)
-        .await?;
-        Ok(rows.into_iter().collect())
+        queries::latest_done_pr_urls_all_projects(&self.pool).await
     }
 
-    /// Count done/failed tasks globally and per project via SQL aggregation.
-    ///
-    /// Returns `(global_done, global_failed, per_project_rows)` where each row is
-    /// `(project_key, done_count, failed_count)`. Tasks with no project are counted
-    /// in the global totals only. Uses the `idx_tasks_project_status_updated` index.
     pub async fn count_done_failed_by_project(
         &self,
     ) -> anyhow::Result<(u64, u64, Vec<(String, u64, u64)>)> {
-        let outcome_statuses = [TaskStatus::Done.as_ref(), TaskStatus::Failed.as_ref()];
-        let global_sql = format!(
-            "SELECT COUNT(CASE WHEN status = 'done' THEN 1 END), \
-                    COUNT(CASE WHEN status = 'failed' THEN 1 END) \
-             FROM tasks WHERE status IN ({})",
-            Self::status_placeholders(outcome_statuses.len())
-        );
-        let global: (i64, i64) = outcome_statuses
-            .iter()
-            .fold(sqlx::query_as(&global_sql), |query, status| {
-                query.bind(*status)
-            })
-            .fetch_one(&self.pool)
-            .await?;
-
-        let rows_sql = format!(
-            "SELECT project, \
-                    COUNT(CASE WHEN status = 'done' THEN 1 END), \
-                    COUNT(CASE WHEN status = 'failed' THEN 1 END) \
-             FROM tasks \
-             WHERE status IN ({}) AND project IS NOT NULL \
-             GROUP BY project",
-            Self::status_placeholders(outcome_statuses.len())
-        );
-        let rows: Vec<(String, i64, i64)> = outcome_statuses
-            .iter()
-            .fold(sqlx::query_as(&rows_sql), |query, status| {
-                query.bind(*status)
-            })
-            .fetch_all(&self.pool)
-            .await?;
-
-        let by_project = rows
-            .into_iter()
-            .map(|(p, d, f)| (p, d as u64, f as u64))
-            .collect();
-        Ok((global.0 as u64, global.1 as u64, by_project))
+        queries::count_done_failed_by_project(&self.pool).await
     }
 
-    /// Return all tasks whose `parent_id` matches the given parent task ID.
     pub async fn list_children(&self, parent_id: &str) -> anyhow::Result<Vec<TaskState>> {
-        let sql = format!(
-            "SELECT {TASK_ROW_COLUMNS} FROM tasks WHERE parent_id = ? ORDER BY created_at DESC"
-        );
-        let rows = sqlx::query_as::<_, TaskRow>(&sql)
-            .bind(parent_id)
-            .fetch_all(&self.pool)
-            .await?;
-        rows.into_iter().map(TaskRow::try_into_task_state).collect()
+        queries::list_children(&self.pool, parent_id).await
     }
 
-    /// Persist a single artifact captured from agent output.
-    ///
-    /// Content larger than [`ARTIFACT_MAX_BYTES`] is truncated to avoid
-    /// unbounded database growth without requiring an external compression
-    /// dependency.
     pub async fn insert_artifact(
         &self,
         task_id: &str,
@@ -782,190 +153,24 @@ impl TaskDb {
         artifact_type: &str,
         content: &str,
     ) -> anyhow::Result<()> {
-        let stored = if content.len() > ARTIFACT_MAX_BYTES {
-            let mut boundary = ARTIFACT_MAX_BYTES;
-            while boundary > 0 && !content.is_char_boundary(boundary) {
-                boundary -= 1;
-            }
-            format!(
-                "{}\n[truncated: {} bytes total]",
-                &content[..boundary],
-                content.len()
-            )
-        } else {
-            content.to_string()
-        };
-
-        sqlx::query(
-            "INSERT INTO task_artifacts (task_id, turn, artifact_type, content)
-             VALUES (?, ?, ?, ?)",
-        )
-        .bind(task_id)
-        .bind(turn as i64)
-        .bind(artifact_type)
-        .bind(&stored)
-        .execute(&self.pool)
-        .await?;
-        Ok(())
+        artifacts::insert_artifact(&self.pool, task_id, turn, artifact_type, content).await
     }
 
-    /// Return all artifacts for a task ordered by insertion time.
     pub async fn list_artifacts(&self, task_id: &str) -> anyhow::Result<Vec<TaskArtifact>> {
-        let rows = sqlx::query_as::<_, TaskArtifact>(
-            "SELECT task_id, turn, artifact_type, content, created_at
-             FROM task_artifacts WHERE task_id = ? ORDER BY id ASC",
-        )
-        .bind(task_id)
-        .fetch_all(&self.pool)
-        .await?;
-        Ok(rows)
+        artifacts::list_artifacts(&self.pool, task_id).await
     }
-}
 
-#[derive(sqlx::FromRow)]
-struct TaskRow {
-    id: String,
-    status: String,
-    turn: i64,
-    pr_url: Option<String>,
-    rounds: String,
-    error: Option<String>,
-    source: Option<String>,
-    external_id: Option<String>,
-    parent_id: Option<String>,
-    created_at: Option<String>,
-    repo: Option<String>,
-    depends_on: String,
-    project: Option<String>,
-    priority: i64,
-    phase: String,
-    description: Option<String>,
-}
-
-impl TaskRow {
-    fn try_into_task_state(self) -> anyhow::Result<TaskState> {
-        let Self {
-            id,
-            status,
-            turn,
-            pr_url,
-            rounds,
-            error,
-            source,
-            external_id,
-            parent_id,
-            created_at,
-            repo,
-            depends_on,
-            project,
-            priority,
-            phase,
-            description,
-        } = self;
-
-        let decoded_rounds = serde_json::from_str(&rounds).map_err(|source| {
-            TaskDbDecodeError::RoundsDeserialize {
-                task_id: id.clone(),
-                source,
-            }
-        })?;
-        let decoded_depends_on = serde_json::from_str(&depends_on).map_err(|source| {
-            TaskDbDecodeError::DependsOnDeserialize {
-                task_id: id.clone(),
-                source,
-            }
-        })?;
-        let decoded_phase =
-            serde_json::from_str::<crate::task_runner::TaskPhase>(&phase).map_err(|source| {
-                TaskDbDecodeError::PhaseDeserialize {
-                    task_id: id.clone(),
-                    source,
-                }
-            })?;
-
-        Ok(TaskState {
-            id: harness_core::types::TaskId(id),
-            status: status.parse::<TaskStatus>()?,
-            turn: turn as u32,
-            pr_url,
-            rounds: decoded_rounds,
-            error,
-            source,
-            external_id,
-            parent_id: parent_id.map(harness_core::types::TaskId),
-            depends_on: decoded_depends_on,
-            subtask_ids: Vec::new(),
-            project_root: project.map(PathBuf::from),
-            issue: None,
-            description,
-            created_at,
-            priority: priority.clamp(0, 255) as u8,
-            phase: decoded_phase,
-            triage_output: None,
-            plan_output: None,
-            repo,
-        })
-    }
-}
-
-/// Lightweight row for summary queries — omits the heavy `rounds` column.
-#[derive(sqlx::FromRow)]
-struct TaskSummaryRow {
-    id: String,
-    status: String,
-    turn: i64,
-    pr_url: Option<String>,
-    error: Option<String>,
-    source: Option<String>,
-    external_id: Option<String>,
-    parent_id: Option<String>,
-    created_at: Option<String>,
-    repo: Option<String>,
-    depends_on: String,
-    project: Option<String>,
-    phase: String,
-    description: Option<String>,
-}
-
-impl TaskSummaryRow {
-    fn try_into_summary(self) -> anyhow::Result<crate::task_runner::TaskSummary> {
-        use harness_core::types::TaskId;
-        let depends_on: Vec<TaskId> = serde_json::from_str(&self.depends_on).map_err(|source| {
-            harness_core::error::TaskDbDecodeError::DependsOnDeserialize {
-                task_id: self.id.clone(),
-                source,
-            }
-        })?;
-        let decoded_phase = serde_json::from_str::<crate::task_runner::TaskPhase>(&self.phase)
-            .map_err(
-                |source| harness_core::error::TaskDbDecodeError::PhaseDeserialize {
-                    task_id: self.id.clone(),
-                    source,
-                },
-            )?;
-        Ok(crate::task_runner::TaskSummary {
-            id: TaskId(self.id),
-            status: self.status.parse::<crate::task_runner::TaskStatus>()?,
-            turn: self.turn as u32,
-            pr_url: self.pr_url,
-            error: self.error,
-            source: self.source,
-            parent_id: self.parent_id.map(TaskId),
-            external_id: self.external_id,
-            repo: self.repo,
-            description: self.description,
-            created_at: self.created_at,
-            phase: decoded_phase,
-            depends_on,
-            subtask_ids: Vec::new(),
-            project: self.project,
-        })
+    // Expose pool for tests that need direct SQL access
+    #[cfg(test)]
+    pub(crate) fn pool(&self) -> &SqlitePool {
+        &self.pool
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::{TaskDb, TaskRow};
+    use super::row::TaskRow;
+    use super::TaskDb;
     use crate::task_runner::{RoundResult, TaskState, TaskStatus};
     use harness_core::error::TaskDbDecodeError;
 
@@ -1032,7 +237,7 @@ mod tests {
             .bind("pending")
             .bind(1_i64)
             .bind("{not-json")
-            .execute(&db.pool)
+            .execute(db.pool())
             .await?;
 
         let err = db
@@ -1056,7 +261,7 @@ mod tests {
         .bind(1_i64)
         .bind("[]")
         .bind("{not-json")
-        .execute(&db.pool)
+        .execute(db.pool())
         .await?;
 
         let err = db
@@ -1703,7 +908,7 @@ mod tests {
         sqlx::query("UPDATE tasks SET status = ? WHERE id = ?")
             .bind("unknown_status")
             .bind("task-unknown")
-            .execute(&db.pool)
+            .execute(db.pool())
             .await?;
 
         let err = db
@@ -1720,7 +925,7 @@ mod tests {
     /// at compile time (struct construction) or at runtime (count mismatch).
     #[test]
     fn task_row_columns_match_struct() {
-        let columns: Vec<&str> = super::TASK_ROW_COLUMNS.split(", ").collect();
+        let columns: Vec<&str> = super::schema::TASK_ROW_COLUMNS.split(", ").collect();
 
         // Build a TaskRow using every column name — compile error if a field is missing.
         let _row = TaskRow {
@@ -1912,7 +1117,7 @@ mod tests {
         .bind(0_i64)
         .bind("[]")
         .bind("[]")
-        .execute(&db.pool)
+        .execute(db.pool())
         .await?;
 
         let loaded = db
