+++ crates/harness-server/src/task_runner/scheduling.rs
@@ -1,0 +1,164 @@
+use super::store::TaskStore;
+use super::types::{CreateTaskRequest, TaskId, TaskState, TaskStatus};
+use std::sync::Arc;
+
+/// Extract `(owner, repo, number)` from a GitHub PR URL.
+///
+/// Expects format: `https://github.com/{owner}/{repo}/pull/{number}[#...]`
+pub(super) fn parse_pr_url(pr_url: &str) -> Option<(String, String, u64)> {
+    // Strip fragment (e.g. #discussion_r...)
+    let url = pr_url.split('#').next().unwrap_or(pr_url);
+    let parts: Vec<&str> = url.trim_end_matches('/').split('/').collect();
+    // Expected: ["https:", "", "github.com", owner, repo, "pull", number]
+    let pull_idx = parts.iter().rposition(|&p| p == "pull")?;
+    if pull_idx + 1 >= parts.len() || pull_idx < 2 {
+        return None;
+    }
+    let number: u64 = parts[pull_idx + 1].parse().ok()?;
+    let repo = parts[pull_idx - 1].to_string();
+    let owner = parts[pull_idx - 2].to_string();
+    Some((owner, repo, number))
+}
+
+/// DFS cycle detection. Returns true if any dependency transitively reaches `new_id`.
+/// Called before the new task is inserted into the store.
+fn detect_cycle(store: &TaskStore, new_id: &TaskId, depends_on: &[TaskId]) -> bool {
+    let mut stack: Vec<TaskId> = depends_on.to_vec();
+    let mut visited: std::collections::HashSet<TaskId> = std::collections::HashSet::new();
+    while let Some(dep_id) = stack.pop() {
+        if dep_id == *new_id {
+            return true;
+        }
+        if visited.contains(&dep_id) {
+            continue;
+        }
+        visited.insert(dep_id.clone());
+        if let Some(dep_state) = store.get(&dep_id) {
+            for transitive in &dep_state.depends_on {
+                stack.push(transitive.clone());
+            }
+        }
+    }
+    false
+}
+
+/// Create a task that waits for its dependencies before starting.
+///
+/// If all deps are already Done, registers as Pending immediately.
+/// If deps are unresolved, registers as AwaitingDeps.
+/// Returns Err if a circular dependency is detected.
+pub async fn spawn_task_awaiting_deps(
+    store: Arc<TaskStore>,
+    req: CreateTaskRequest,
+) -> anyhow::Result<TaskId> {
+    let depends_on = req.depends_on.clone();
+    let task_id = TaskId::new();
+
+    if detect_cycle(&store, &task_id, &depends_on) {
+        anyhow::bail!("circular dependency detected for task {}", task_id.0);
+    }
+
+    let mut all_done = true;
+    for dep_id in &depends_on {
+        if !matches!(store.dep_status(dep_id).await, Some(TaskStatus::Done)) {
+            all_done = false;
+            break;
+        }
+    }
+
+    let mut state = TaskState::new(task_id.clone());
+    state.depends_on = depends_on;
+    state.source = req.source.clone();
+    state.external_id = req.external_id.clone();
+    state.repo = req.repo.clone();
+    state.priority = req.priority;
+    if let Some(parent) = req.parent_task_id.clone() {
+        state.parent_id = Some(parent);
+    }
+
+    if !all_done && !state.depends_on.is_empty() {
+        state.status = TaskStatus::AwaitingDeps;
+    }
+
+    store.insert(&state).await;
+    store.register_task_stream(&task_id);
+    Ok(task_id)
+}
+
+/// Check all AwaitingDeps tasks and transition ready ones to Pending.
+/// Returns the IDs of tasks that were transitioned to Pending.
+///
+/// Async because dependency status checks fall back to the database for
+/// terminal tasks (Done/Failed) that were evicted from the startup cache.
+///
+/// Returns `(ready_ids, newly_failed_ids)`.  Both sets must be persisted by
+/// the caller so that status transitions survive a process restart.
+pub async fn check_awaiting_deps(store: &TaskStore) -> (Vec<TaskId>, Vec<TaskId>) {
+    let mut ready = Vec::new();
+    let mut failed_deps: Vec<(TaskId, TaskId)> = Vec::new();
+
+    // Snapshot AwaitingDeps tasks from cache before async work.
+    let awaiting: Vec<(TaskId, Vec<TaskId>)> = store
+        .cache
+        .iter()
+        .filter_map(|e| {
+            let task = e.value();
+            if matches!(task.status, TaskStatus::AwaitingDeps) {
+                Some((task.id.clone(), task.depends_on.clone()))
+            } else {
+                None
+            }
+        })
+        .collect();
+
+    for (task_id, depends_on) in awaiting {
+        // Single pass: one DB lookup per dep (cache-first, then lightweight
+        // `SELECT status` — no `rounds` JSON decode).
+        let mut failed_dep_id = None;
+        let mut all_done = true;
+        for dep_id in &depends_on {
+            match store.dep_status(dep_id).await {
+                Some(TaskStatus::Failed) => {
+                    failed_dep_id = Some(dep_id.clone());
+                    break; // fail-fast — no need to inspect remaining deps
+                }
+                Some(TaskStatus::Done) => {} // this dep is satisfied
+                _ => {
+                    // Pending, Implementing, or unknown — not ready yet.
+                    // Keep scanning in case a later dep is Failed.
+                    all_done = false;
+                }
+            }
+        }
+        if let Some(fd) = failed_dep_id {
+            failed_deps.push((task_id, fd));
+            continue;
+        }
+        if all_done {
+            ready.push(task_id);
+        }
+    }
+
+    let mut newly_failed: Vec<TaskId> = Vec::new();
+    for (task_id, failed_dep_id) in &failed_deps {
+        if let Some(mut entry) = store.cache.get_mut(task_id) {
+            // Only overwrite if the task is still AwaitingDeps — a concurrent
+            // cancel or status transition must not be clobbered.
+            if matches!(entry.status, TaskStatus::AwaitingDeps) {
+                entry.status = TaskStatus::Failed;
+                entry.error = Some(format!("dependency {} failed", failed_dep_id.0));
+                newly_failed.push(task_id.clone());
+            }
+        }
+    }
+    for task_id in &ready {
+        if let Some(mut entry) = store.cache.get_mut(task_id) {
+            // Same guard: skip if status changed since we snapshotted.
+            if matches!(entry.status, TaskStatus::AwaitingDeps) {
+                entry.status = TaskStatus::Pending;
+            }
+        }
+    }
+
+    (ready, newly_failed)
+}
