+++ crates/harness-server/src/task_runner/types.rs
@@ -1,0 +1,598 @@
+use harness_core::types::TaskId as CoreTaskId;
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::future::Future;
+use std::path::PathBuf;
+use std::pin::Pin;
+use std::sync::Arc;
+
+pub type TaskId = CoreTaskId;
+
+/// Broadcast channel capacity for per-task stream events.
+/// When the buffer is full, the oldest events are dropped for lagging receivers.
+pub(crate) const TASK_STREAM_CAPACITY: usize = 512;
+
+/// Async callback invoked when a task reaches a terminal state (Done/Failed).
+/// Receives a snapshot of the final `TaskState`. Implemented in the intake layer
+/// to call `IntakeSource::on_task_complete` without creating circular dependencies.
+pub type CompletionCallback =
+    Arc<dyn Fn(TaskState) -> Pin<Box<dyn Future<Output = ()> + Send>> + Send + Sync>;
+
+/// Current phase in the task pipeline.
+///
+/// Tasks progress through phases sequentially. Simple tasks may skip
+/// Triage/Plan and go directly to Implement.
+#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum TaskPhase {
+    /// Initial phase — Tech Lead evaluates the issue.
+    Triage,
+    /// Architect designs the implementation plan.
+    Plan,
+    /// Engineer writes and tests code (default starting phase).
+    #[default]
+    Implement,
+    /// Independent code review.
+    Review,
+    /// Terminal — task completed or failed.
+    Terminal,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum TaskStatus {
+    Pending,
+    AwaitingDeps,
+    Implementing,
+    AgentReview,
+    Waiting,
+    Reviewing,
+    Done,
+    Failed,
+    Cancelled,
+}
+
+const TERMINAL_TASK_STATUSES: &[&str] = &["done", "failed", "cancelled"];
+const RESUMABLE_TASK_STATUSES: &[&str] = &["implementing", "agent_review", "waiting", "reviewing"];
+
+impl TaskStatus {
+    pub fn is_terminal(&self) -> bool {
+        matches!(self, Self::Done | Self::Failed | Self::Cancelled)
+    }
+
+    pub fn is_inflight(&self) -> bool {
+        matches!(
+            self,
+            Self::Implementing | Self::AgentReview | Self::Waiting | Self::Reviewing
+        )
+    }
+
+    pub fn is_resumable_after_restart(&self) -> bool {
+        self.is_inflight()
+    }
+
+    pub fn is_success(&self) -> bool {
+        matches!(self, Self::Done)
+    }
+
+    pub fn is_failure(&self) -> bool {
+        matches!(self, Self::Failed)
+    }
+
+    pub fn is_cancelled(&self) -> bool {
+        matches!(self, Self::Cancelled)
+    }
+
+    pub fn terminal_statuses() -> &'static [&'static str] {
+        TERMINAL_TASK_STATUSES
+    }
+
+    pub fn resumable_statuses() -> &'static [&'static str] {
+        RESUMABLE_TASK_STATUSES
+    }
+}
+
+impl AsRef<str> for TaskStatus {
+    fn as_ref(&self) -> &str {
+        match self {
+            TaskStatus::Pending => "pending",
+            TaskStatus::AwaitingDeps => "awaiting_deps",
+            TaskStatus::Implementing => "implementing",
+            TaskStatus::AgentReview => "agent_review",
+            TaskStatus::Waiting => "waiting",
+            TaskStatus::Reviewing => "reviewing",
+            TaskStatus::Done => "done",
+            TaskStatus::Failed => "failed",
+            TaskStatus::Cancelled => "cancelled",
+        }
+    }
+}
+
+impl std::str::FromStr for TaskStatus {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "pending" => Ok(TaskStatus::Pending),
+            "awaiting_deps" => Ok(TaskStatus::AwaitingDeps),
+            "implementing" => Ok(TaskStatus::Implementing),
+            "agent_review" => Ok(TaskStatus::AgentReview),
+            "waiting" => Ok(TaskStatus::Waiting),
+            "reviewing" => Ok(TaskStatus::Reviewing),
+            "done" => Ok(TaskStatus::Done),
+            "failed" => Ok(TaskStatus::Failed),
+            "cancelled" => Ok(TaskStatus::Cancelled),
+            _ => anyhow::bail!("unknown task status `{s}`"),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct RoundResult {
+    pub turn: u32,
+    pub action: String,
+    pub result: String,
+    /// Raw output from the reviewer agent, if available.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub detail: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct TaskState {
+    pub id: TaskId,
+    pub status: TaskStatus,
+    pub turn: u32,
+    pub pr_url: Option<String>,
+    pub rounds: Vec<RoundResult>,
+    pub error: Option<String>,
+    /// Intake source name that created this task (e.g. "github", "feishu").
+    /// Persisted so the association survives server restart.
+    pub source: Option<String>,
+    /// Source-specific identifier for the originating issue/message.
+    /// Used by `CompletionCallback` to call `IntakeSource::on_task_complete`.
+    pub external_id: Option<String>,
+    /// Parent task ID when this task is a subtask spawned via parallel dispatch.
+    /// Persisted so parent-child relationships survive server restart.
+    pub parent_id: Option<TaskId>,
+    /// Task IDs that must reach Done before this task may start.
+    /// Persisted as JSON in the database.
+    #[serde(default)]
+    pub depends_on: Vec<TaskId>,
+    /// IDs of subtasks spawned by this task during parallel dispatch.
+    /// Populated at runtime; not persisted (use `TaskStore::list_children` after restart).
+    #[serde(default)]
+    pub subtask_ids: Vec<TaskId>,
+    /// Resolved project root for this task. Set at spawn time and persisted to the database.
+    /// Used by sibling-awareness lookups and exposed via TaskSummary for observability.
+    #[serde(skip)]
+    pub project_root: Option<PathBuf>,
+    /// GitHub issue number if this is an issue-based task. Set at spawn time; not persisted.
+    #[serde(skip)]
+    pub issue: Option<u64>,
+    /// Repository slug (e.g. "owner/repo"). Persisted for traceability.
+    pub repo: Option<String>,
+    /// Short description derived from the task prompt or issue number.
+    #[serde(default)]
+    pub description: Option<String>,
+    /// ISO 8601 creation timestamp. Set at spawn time and persisted to the tasks DB.
+    #[serde(default)]
+    pub created_at: Option<String>,
+    /// Scheduling priority: 0 = normal (default), 1 = high, 2 = critical.
+    /// Persisted to DB and used by TaskQueue::acquire to skip lower-priority waiters.
+    #[serde(default)]
+    pub priority: u8,
+    /// Current pipeline phase. Defaults to Implement for backward compatibility.
+    #[serde(default)]
+    pub phase: TaskPhase,
+    /// Output from the Triage phase (Tech Lead assessment). Not persisted to DB.
+    #[serde(skip)]
+    pub triage_output: Option<String>,
+    /// Output from the Plan phase (Architect plan). Not persisted to DB.
+    #[serde(skip)]
+    pub plan_output: Option<String>,
+}
+
+/// Lightweight task summary returned by the list endpoint (excludes `rounds` history).
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct TaskSummary {
+    pub id: TaskId,
+    pub status: TaskStatus,
+    pub turn: u32,
+    pub pr_url: Option<String>,
+    pub error: Option<String>,
+    /// Intake source name (e.g. "github", "feishu", "dashboard"). None for manual tasks.
+    pub source: Option<String>,
+    /// Parent task ID when this is a subtask.
+    pub parent_id: Option<TaskId>,
+    /// Source-specific identifier (e.g. GitHub issue number).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub external_id: Option<String>,
+    /// Repository slug (e.g. "owner/repo").
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub repo: Option<String>,
+    /// Short description derived from the task prompt or issue number.
+    #[serde(default)]
+    pub description: Option<String>,
+    /// ISO 8601 creation timestamp.
+    #[serde(default)]
+    pub created_at: Option<String>,
+    /// Current pipeline phase.
+    #[serde(default)]
+    pub phase: TaskPhase,
+    /// Task IDs that must reach Done before this task may start.
+    #[serde(default)]
+    pub depends_on: Vec<TaskId>,
+    /// IDs of subtasks spawned by this task during parallel dispatch.
+    #[serde(default)]
+    pub subtask_ids: Vec<TaskId>,
+    /// Resolved project root path. Persisted to the database for observability.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub project: Option<String>,
+}
+
+impl TaskState {
+    pub(crate) fn new(id: TaskId) -> Self {
+        Self {
+            id,
+            status: TaskStatus::Pending,
+            turn: 0,
+            pr_url: None,
+            rounds: Vec::new(),
+            error: None,
+            source: None,
+            external_id: None,
+            parent_id: None,
+            depends_on: Vec::new(),
+            subtask_ids: Vec::new(),
+            project_root: None,
+            issue: None,
+            description: None,
+            created_at: Some(chrono::Utc::now().to_rfc3339()),
+            priority: 0,
+            phase: TaskPhase::default(),
+            triage_output: None,
+            plan_output: None,
+            repo: None,
+        }
+    }
+
+    pub fn summary(&self) -> TaskSummary {
+        TaskSummary {
+            id: self.id.clone(),
+            status: self.status.clone(),
+            turn: self.turn,
+            pr_url: self.pr_url.clone(),
+            error: self.error.clone(),
+            source: self.source.clone(),
+            parent_id: self.parent_id.clone(),
+            external_id: self.external_id.clone(),
+            repo: self.repo.clone(),
+            description: self.description.clone(),
+            created_at: self.created_at.clone(),
+            phase: self.phase.clone(),
+            depends_on: self.depends_on.clone(),
+            subtask_ids: self.subtask_ids.clone(),
+            project: self
+                .project_root
+                .as_ref()
+                .map(|p| p.to_string_lossy().into_owned()),
+        }
+    }
+}
+
+/// Maximum allowed scheduling priority. Values above this are rejected at the
+/// API boundary to prevent scheduler-level starvation of normal-priority tasks.
+/// `0` = normal (default), `1` = high, `2` = critical.
+pub const MAX_TASK_PRIORITY: u8 = 2;
+
+#[derive(Debug, Clone, Deserialize)]
+pub struct CreateTaskRequest {
+    /// Free-text task description (prompt, issue URL, etc.).
+    pub prompt: Option<String>,
+    /// GitHub issue number to implement from.
+    pub issue: Option<u64>,
+    /// GitHub PR number to review/fix.
+    pub pr: Option<u64>,
+    /// Explicit agent name; if omitted, uses the default agent.
+    pub agent: Option<String>,
+    /// Project root; defaults to the main git worktree resolved at task spawn time.
+    pub project: Option<PathBuf>,
+    #[serde(default = "default_wait")]
+    pub wait_secs: u64,
+    /// Maximum review rounds. When absent, triage complexity provides the default
+    /// (Low=2, Medium=system default, High=8). When explicitly set by the caller,
+    /// this value always wins over triage-derived defaults.
+    #[serde(default)]
+    pub max_rounds: Option<u32>,
+    /// Maximum total agent API calls across all phases (implementation + validation retries +
+    /// review rounds). When None, falls back to the global `concurrency.max_turns` config value.
+    #[serde(default)]
+    pub max_turns: Option<u32>,
+    /// Per-turn timeout in seconds; defaults to 3600 (1 hour).
+    #[serde(default = "default_turn_timeout")]
+    pub turn_timeout_secs: u64,
+    /// Maximum spend budget for the agent in USD; None means unlimited.
+    #[serde(default)]
+    pub max_budget_usd: Option<f64>,
+    /// Base delay in milliseconds for the first validation retry; subsequent retries double the
+    /// delay up to `retry_max_backoff_ms`. Default: 10 000 ms (10 s).
+    #[serde(default = "default_retry_base_backoff_ms")]
+    pub retry_base_backoff_ms: u64,
+    /// Maximum backoff cap in milliseconds for validation retries. Default: 300 000 ms (5 min).
+    #[serde(default = "default_retry_max_backoff_ms")]
+    pub retry_max_backoff_ms: u64,
+    /// Seconds of silence from the agent stream before declaring a stall; defaults to 300.
+    /// Overrides the global `concurrency.stall_timeout_secs` for this task.
+    #[serde(default = "default_stall_timeout")]
+    pub stall_timeout_secs: u64,
+    /// Intake source name (e.g. "github", "feishu", "periodic_review"). None for manual tasks.
+    #[serde(default)]
+    pub source: Option<String>,
+    /// Source-specific identifier (e.g. GitHub issue number). Stored in TaskState for traceability.
+    #[serde(default)]
+    pub external_id: Option<String>,
+    /// Repository slug (e.g. "owner/repo"). Stored in TaskState for traceability.
+    #[serde(default)]
+    pub repo: Option<String>,
+    /// Explicit parent task ID.
+    #[serde(default)]
+    pub parent_task_id: Option<TaskId>,
+    /// Task IDs that must complete (Done) before this task starts.
+    #[serde(default)]
+    pub depends_on: Vec<TaskId>,
+    /// Scheduling priority: 0 = normal (default), 1 = high, 2 = critical.
+    /// Higher values are served first when multiple tasks are waiting for a slot.
+    #[serde(default)]
+    pub priority: u8,
+}
+
+impl Default for CreateTaskRequest {
+    fn default() -> Self {
+        Self {
+            prompt: None,
+            issue: None,
+            pr: None,
+            agent: None,
+            project: None,
+            wait_secs: default_wait(),
+            max_rounds: None,
+            max_turns: None,
+            turn_timeout_secs: default_turn_timeout(),
+            max_budget_usd: None,
+            retry_base_backoff_ms: default_retry_base_backoff_ms(),
+            retry_max_backoff_ms: default_retry_max_backoff_ms(),
+            stall_timeout_secs: default_stall_timeout(),
+            source: None,
+            external_id: None,
+            repo: None,
+            parent_task_id: None,
+            depends_on: Vec::new(),
+            priority: 0,
+        }
+    }
+}
+
+pub(crate) fn summarize_request_description(req: &CreateTaskRequest) -> Option<String> {
+    // Only persist structured safe labels — never raw prompt text, which may contain
+    // credentials or customer data.
+    if let Some(n) = req.issue {
+        return Some(format!("issue #{n}"));
+    }
+    if let Some(n) = req.pr {
+        return Some(format!("PR #{n}"));
+    }
+    // Prompt-only tasks: store a generic label so that:
+    //   (a) sibling-awareness can include them (prevents parallel agents stomping the same files),
+    //   (b) operators can identify crashed tasks in the DB/dashboard after a restart.
+    // The prompt itself is deliberately not stored.
+    if req.prompt.is_some() {
+        return Some("prompt task".to_string());
+    }
+    None
+}
+
+pub(crate) async fn fill_missing_repo_from_project(req: &mut CreateTaskRequest) {
+    if req.repo.is_some() {
+        return;
+    }
+    let Some(project) = req.project.as_deref() else {
+        return;
+    };
+    req.repo = crate::task_executor::pr_detection::detect_repo_slug(project).await;
+}
+
+/// Detect the main git worktree root using a blocking subprocess call.
+/// Must be called via `tokio::task::spawn_blocking` in async contexts.
+pub(super) fn detect_main_worktree() -> PathBuf {
+    std::process::Command::new("git")
+        .args(["worktree", "list", "--porcelain"])
+        .output()
+        .ok()
+        .and_then(|o| String::from_utf8(o.stdout).ok())
+        .and_then(|s| {
+            s.lines()
+                .next()
+                .and_then(|line| line.strip_prefix("worktree "))
+                .map(|p| PathBuf::from(p.trim()))
+        })
+        .unwrap_or_else(|| {
+            tracing::warn!(
+                "detect_main_worktree: could not detect git worktree root, falling back to '.'"
+            );
+            PathBuf::from(".")
+        })
+}
+
+pub(super) fn default_wait() -> u64 {
+    120
+}
+
+pub(super) fn default_retry_base_backoff_ms() -> u64 {
+    10_000
+}
+pub(super) fn default_retry_max_backoff_ms() -> u64 {
+    300_000
+}
+pub(super) fn default_turn_timeout() -> u64 {
+    // 1 hour: parallel subtasks and complex agent turns on large codebases
+    // regularly exceed the previous 10-minute default when running CI checks,
+    // building dependencies from source, or iterating on review feedback.
+    3600
+}
+
+/// Convert a user-facing timeout value to a Duration.
+/// `0` means "no timeout" and maps to ~277 hours (effectively unlimited).
+pub(crate) fn effective_turn_timeout(secs: u64) -> tokio::time::Duration {
+    if secs == 0 {
+        tokio::time::Duration::from_secs(999_999)
+    } else {
+        tokio::time::Duration::from_secs(secs)
+    }
+}
+pub(super) fn default_stall_timeout() -> u64 {
+    300
+}
+
+pub(super) fn describe_detect_main_worktree_join_error(
+    join_err: &tokio::task::JoinError,
+) -> String {
+    if join_err.is_panic() {
+        format!("detect_main_worktree panicked: {join_err}")
+    } else if join_err.is_cancelled() {
+        format!("detect_main_worktree was cancelled: {join_err}")
+    } else {
+        format!("detect_main_worktree failed: {join_err}")
+    }
+}
+
+pub(crate) async fn resolve_project_root_with(
+    requested_project: Option<PathBuf>,
+    detect_worktree: impl FnOnce() -> PathBuf + Send + 'static,
+) -> anyhow::Result<PathBuf> {
+    match requested_project {
+        Some(project) => Ok(project),
+        None => tokio::task::spawn_blocking(detect_worktree)
+            .await
+            .map_err(|join_err| {
+                let reason = describe_detect_main_worktree_join_error(&join_err);
+                tracing::error!("{reason}");
+                anyhow::anyhow!("{reason}")
+            }),
+    }
+}
+
+/// Resolve and canonicalize the project root so the caller can obtain a stable
+/// semaphore key before acquiring the concurrency permit.
+///
+/// When `project` is `None` the main git worktree is detected automatically.
+/// Symlinks, relative paths and the `None` sentinel all converge to the same
+/// canonical `PathBuf`, preventing the same repository from landing in
+/// different per-project buckets due to path aliasing.
+///
+/// This function only resolves symlinks — it does NOT enforce the HOME-boundary
+/// restriction applied by `validate_project_root`. Full validation still
+/// happens inside `spawn_task` once the task is running.
+pub(crate) async fn resolve_canonical_project(project: Option<PathBuf>) -> anyhow::Result<PathBuf> {
+    let raw = resolve_project_root_with(project, detect_main_worktree).await?;
+    // Best-effort canonicalize: if the path doesn't exist yet (e.g. in tests
+    // using a path that will be created later) fall back to the raw path so
+    // we at least get a consistent string key.
+    Ok(raw.canonicalize().unwrap_or(raw))
+}
+
+/// Patterns that indicate a transient (retryable) failure rather than a permanent one.
+pub(super) const TRANSIENT_PATTERNS: &[&str] = &[
+    "at capacity",
+    "rate limit",
+    "rate_limit",
+    "hit your limit",
+    "429",
+    "502 Bad Gateway",
+    "503 Service",
+    "overloaded",
+    "connection reset",
+    "connection refused",
+    "broken pipe",
+    "EOF",
+    "stream idle timeout",
+    "stream stall",
+    "ECONNRESET",
+    "ETIMEDOUT",
+    // SQLite transient contention — SQLITE_BUSY / SQLITE_LOCKED
+    "database is locked",
+    "database table is locked",
+    "SQLITE_BUSY",
+    "SQLITE_LOCKED",
+];
+
+/// Pattern indicating the CLI account-level usage limit has been reached.
+/// When detected, the global rate-limit circuit breaker is activated to
+/// prevent all other tasks from wasting turns.
+pub(super) const ACCOUNT_LIMIT_PATTERN: &str = "hit your limit";
+
+/// Maximum number of automatic retries for transient failures.
+pub(super) const MAX_TRANSIENT_RETRIES: u32 = 2;
+
+/// Return `true` when a free-text prompt is complex enough to require a Plan phase
+/// before implementation.
+///
+/// Heuristic: prompt longer than 200 words OR contains 3 or more file-path-like
+/// tokens (sequences containing `/` or ending in a recognised source extension).
+pub(crate) fn prompt_requires_plan(prompt: &str) -> bool {
+    let word_count = prompt.split_whitespace().count();
+    if word_count > 200 {
+        return true;
+    }
+    let file_path_count = prompt
+        .split_whitespace()
+        .filter(|tok| {
+            // Exclude XML/HTML tags — `</foo>` contains '/' but is not a file path.
+            // System prompts wrap user data in `<external_data>...</external_data>`
+            // which would otherwise trigger false positives.
+            let is_xml_tag = tok.starts_with('<');
+            if is_xml_tag {
+                return false;
+            }
+            tok.contains('/') || {
+                let lower = tok.to_lowercase();
+                lower.ends_with(".rs")
+                    || lower.ends_with(".ts")
+                    || lower.ends_with(".tsx")
+                    || lower.ends_with(".go")
+                    || lower.ends_with(".py")
+                    || lower.ends_with(".toml")
+                    || lower.ends_with(".json")
+            }
+        })
+        .count();
+    file_path_count >= 3
+}
+
+pub(crate) fn is_non_decomposable_prompt_source(source: Option<&str>) -> bool {
+    matches!(source, Some("periodic_review") | Some("sprint_planner"))
+}
+
+/// Check if an error message indicates a transient failure that may succeed on retry.
+pub(crate) fn is_transient_error(reason: &str) -> bool {
+    let lower = reason.to_lowercase();
+    TRANSIENT_PATTERNS
+        .iter()
+        .any(|p| lower.contains(&p.to_lowercase()))
+}
+
+/// In-memory cache + SQLite persistence.
+/// Per-project done/failed task counts derived from the in-memory cache.
+#[derive(Debug, Default, Clone)]
+pub struct ProjectCounts {
+    pub done: u64,
+    pub failed: u64,
+}
+
+/// Combined global and per-project done/failed counts produced by a single
+/// cache scan, avoiding both full task cloning and double iteration.
+#[derive(Debug)]
+pub struct DashboardCounts {
+    pub global_done: u64,
+    pub global_failed: u64,
+    pub by_project: HashMap<String, ProjectCounts>,
+}
